[DOTFILES] Remove TODO from README
The biggest difference so far was the git identity which was solved by
using conditional include.

The other difference is using *nux vs macOS but let's be honest I have
not start up my personal laptop in months and when I work on personal
"projects" I do it from my work laptop. Plus it's always possible for
the bashrc to be compatible with both OS.